Tested code -no improvement
The new external interrupts didn't appear to improve accuracy
with the motor. It's not clear to me if the interrupt routine
is missing counts, or if I've screwed up the math somewhere
along the line. It appear to be missing 20-40 triggers, so that
even a short 10-15 count movement is turning an extra 16-32
counts (without registering those counts in the total).