Implemented new logic tests for moving motor
The decision to move the motor now relies on checking
that the new target position is within both the lower
and upper limits of the rack, and that the appropriate
limit switch hasn't already been activated.

Also modified the conversion factor for distance per
motor step to work with the 10tpi lead screw.