New version with external tide calculation library
Moved all of the tide calculation portions to an
external library to simplify the appearance of the
code in this main program. This simplifies the process
of switching between tide sites as well, since you only
need to include a new library for the new site rather than
changing many lines of code inside this program.