Implemented ENABLE line on stepper driver
The Big Easy Driver has an enable line that can be used to
shut down power to the stepper motor when pulled high.
Because the lead-screw design of the rack holds position when
the motor is powered off, I can use the enable line to shut
down the motor between movements and thereby reduce the heat
production of the Big Easy Driver and the motor. It has the
added side effect of reducing the whining noise that came from
the driver at idle. This cuts the idle temperature of the driver
chip from approx 75C to 35C, since it's off most of the time.