Adding helpful comments.
Altered code for lowerPos so that it simply calculates the value
based on the user's chosen upperPos value. There is a hard-coded
value of 2.75 in there to represent the travel limits of the
first tide control rack that I built. The goal is to simplify things
so that the user only has to set one value (upperPos) if they want
to alter the tide range that the tide rack covers.